refactor: migrate search parameters to config directory
- Moved search parameters and related functions from the lib directory to a new config directory, enhancing code organization and maintainability.
- Updated import paths in multiple files to reflect this change.
